+++ src/app/mypage/page.tsx
@@ -2,187 +2,45 @@ import { redirect } from "next/navigation";
 import Image from "next/image";
 import Link from "next/link";
 import { getCurrentUser } from "@/lib/auth";
-import { getDb } from "@/lib/db";
 import { getDbProvider } from "@/lib/env";
-import { characters } from "@/data/characters";
-import { DeckManager } from "@/services/tarot/deck-manager";
 import { getRequestLocale } from "@/i18n/server-locale";
-import { getCardName } from "@/data/cards/locale-helpers";
 import { t } from "@/i18n/translations";
-import type { Locale } from "@/i18n/config";
 import { getTopicLabel } from "@/data/topics-meta";
 import { LogoutButton } from "./LogoutButton";
 import { FavoriteCharacterSelector } from "./FavoriteCharacterSelector";
-
-interface ReadingData {
-  id: string;
-  session_id?: string;
-  share_token?: string | null;
-  overall_reading?: string | null;
-}
-
-interface SessionRow {
-  id: string;
-  service_type: string;
-  topic: string;
-  status: string;
-  created_at: string;
-  character_id?: string | null;
-  readings: ReadingData | ReadingData[] | null;
-  saju_readings: ReadingData | ReadingData[] | null;
-  shinjeom_readings: ReadingData | ReadingData[] | null;
-}
-
-interface SessionCard {
-  card_id: string;
-}
-
-interface Profile {
-  id: string;
-  email?: string;
-  nickname?: string;
-  avatar_url?: string;
-  provider?: string;
-  favorite_character_id?: string;
-}
-
-const topicColors: Record<string, string> = {
-  love: "bg-pink-500/20 text-pink-300 border-pink-500/30",
-  "love-single": "bg-pink-500/20 text-pink-300 border-pink-500/30",
-  "love-couple": "bg-rose-500/20 text-rose-300 border-rose-500/30",
-  finance: "bg-yellow-500/20 text-yellow-300 border-yellow-500/30",
-  career: "bg-blue-500/20 text-blue-300 border-blue-500/30",
-  health: "bg-green-500/20 text-green-300 border-green-500/30",
-  general: "bg-arcana-purple/20 text-arcana-purple border-arcana-purple/30",
-};
-
-const serviceColors: Record<string, string> = {
-  tarot: "text-arcana-purple",
-  saju: "text-cyan-400",
-  shinjeom: "text-amber-400",
-};
-
-const deckManager = new DeckManager();
-
-function formatRelativeDate(dateStr: string, locale: Locale): string {
-  const date = new Date(dateStr);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-  if (diffDays === 0) return t("mypage.date.today", locale);
-  if (diffDays === 1) return t("mypage.date.yesterday", locale);
-  if (diffDays < 7) return t("mypage.date.days-ago", locale).replace("{n}", String(diffDays));
-  if (diffDays < 30) return t("mypage.date.weeks-ago", locale).replace("{n}", String(Math.floor(diffDays / 7)));
-  // 월 단위 이상은 locale 별 toLocaleDateString
-  const localeMap: Record<Locale, string> = { ko: "ko-KR", en: "en-US", ja: "ja-JP" };
-  return date.toLocaleDateString(localeMap[locale]);
-}
-
-function getCharacterName(characterId?: string | null): string | null {
-  if (!characterId) return null;
-  const char = characters.find((c) => c.id === characterId);
-  return char?.name ?? null;
-}
-
-/** PostgREST 1:1 관계: 객체 또는 배열 모두 처리 */
-function normalizeReading(readings: ReadingData | ReadingData[] | null | undefined): ReadingData | undefined {
-  if (!readings) return undefined;
-  if (Array.isArray(readings)) return readings[0];
-  return readings;
-}
-
-/** 3개 리딩 테이블 중 데이터가 있는 것 반환 */
-function getReadingFromSession(session: SessionRow): ReadingData | undefined {
-  return normalizeReading(session.readings)
-    ?? normalizeReading(session.saju_readings)
-    ?? normalizeReading(session.shinjeom_readings);
-}
-
-/** 카드 빈도 집계 → 가장 많이 뽑은 카드명 반환 */
-function getMostFrequentCard(sessionCards: SessionCard[], locale: Locale): string | null {
-  if (sessionCards.length === 0) return null;
-  const freq: Record<string, number> = {};
-  for (const sc of sessionCards) {
-    freq[sc.card_id] = (freq[sc.card_id] || 0) + 1;
-  }
-  const topCardId = Object.entries(freq).sort((a, b) => b[1] - a[1])[0][0];
-  const card = deckManager.getCardById(topCardId);
-  return card ? getCardName(card, locale) : topCardId;
-}
-
-function getServiceLabel(serviceType: string, locale: Locale): string {
-  if (serviceType === "saju") return t("mypage.service.saju", locale);
-  if (serviceType === "shinjeom") return t("mypage.service.shinjeom", locale);
-  return t("mypage.service.tarot", locale);
-}
+import { fetchMypageData } from "./mypage-queries";
+import {
+  formatRelativeDate,
+  getCharacterName,
+  getReadingFromSession,
+  getServiceLabel,
+  topicColors,
+  serviceColors,
+} from "./mypage-utils";
 
 export default async function MyPage() {
   const user = await getCurrentUser();
   if (!user) redirect("/auth/login");
 
   const locale = await getRequestLocale();
-  const db = getDb();
 
-  // 1. 프로필 + 전체 세션 병렬 조회
-  const [profile, allRawSessions] = await Promise.all([
-    db.findOne<Profile>("profiles", { id: user.id }).catch(() => null),
-    db.findMany<SessionRow>("sessions", { user_id: user.id }).catch(() => [] as SessionRow[]),
-  ]);
-
-  const profileError: { message: string } | null = !profile ? { message: t("mypage.profile.error-detail-fallback", locale) } : null;
-
-  // status 필터 + 최신순 정렬 (DB Provider별 ORDER BY 미지원이므로 JS에서 처리)
-  const filtered = (allRawSessions as SessionRow[])
-    .filter((s) => s.status === "completed" || s.status === "in_progress")
-    .sort((a, b) => new Date(b.created_at).getTime() - new Date(a.created_at).getTime())
-    .slice(0, 50);
-
-  // 2. 서비스 타입별 ID 분류 → 3개 테이블 일괄 조회 (N+1 → 3 queries)
-  const tarotIds = filtered.filter((s) => s.service_type === "tarot").map((s) => s.id);
-  const sajuIds = filtered.filter((s) => s.service_type === "saju").map((s) => s.id);
-  const shinjeomIds = filtered.filter((s) => s.service_type === "shinjeom").map((s) => s.id);
-
-  const [tarotReadings, sajuReadings, shinjeomReadings] = await Promise.all([
-    db.findManyIn<ReadingData>("readings", "session_id", tarotIds).catch(() => [] as ReadingData[]),
-    db.findManyIn<ReadingData>("saju_readings", "session_id", sajuIds).catch(() => [] as ReadingData[]),
-    db.findManyIn<ReadingData>("shinjeom_readings", "session_id", shinjeomIds).catch(() => [] as ReadingData[]),
-  ]);
-
-  const tarotMap = new Map(tarotReadings.map((r) => [r.session_id, r]));
-  const sajuMap = new Map(sajuReadings.map((r) => [r.session_id, r]));
-  const shinjeomMap = new Map(shinjeomReadings.map((r) => [r.session_id, r]));
-
-  const withReadings = filtered.map((session) => ({
-    ...session,
-    readings: session.service_type === "tarot" ? (tarotMap.get(session.id) ?? null) : null,
-    saju_readings: session.service_type === "saju" ? (sajuMap.get(session.id) ?? null) : null,
-    shinjeom_readings: session.service_type === "shinjeom" ? (shinjeomMap.get(session.id) ?? null) : null,
-  })) as SessionRow[];
-
-  // completed 세션 + 리딩이 있는 in_progress 세션 (상태 업데이트 누락 복구)
-  const sessionList = withReadings.filter((s) => {
-    if (s.status === "completed") return true;
-    return !!getReadingFromSession(s);
-  }).slice(0, 20);
-
-  // 3. session_cards 일괄 조회 (N+1 → 1 query)
-  const sessionIds = sessionList.map((s) => s.id);
-  const sessionCards: SessionCard[] = sessionIds.length > 0
-    ? await db.findManyIn<SessionCard>("session_cards", "session_id", sessionIds).catch(() => [])
-    : [];
-
-  const totalReadings = sessionList.length;
-  const lastReadingDate = sessionList.length > 0 ? sessionList[0].created_at : null;
-  const favoriteCharName = getCharacterName(profile?.favorite_character_id);
-  const mostFrequentCard = getMostFrequentCard(sessionCards, locale);
-  const nickname = profile?.nickname || t("mypage.profile.default-nickname", locale);
-  const initial = nickname.charAt(0);
+  const {
+    profile,
+    profileError,
+    sessionList,
+    totalReadings,
+    lastReadingDate,
+    favoriteCharName,
+    mostFrequentCard,
+    nickname,
+    initial,
+  } = await fetchMypageData(user.id, locale);
 
   return (
     <div className="relative min-h-screen overflow-hidden">
       {/* 배경 이미지 */}
       <div className="fixed inset-0 -z-10">
-        <Image src="/images/backgrounds/mypage-bg.jpg" alt="" fill className="object-cover"  sizes="100vw" />
+        <Image src="/images/backgrounds/mypage-bg.jpg" alt="" fill className="object-cover" sizes="100vw" />
         <div className="absolute inset-0 bg-arcana-bg/60" />
         <div className="absolute inset-0 bg-[radial-gradient(ellipse_at_center,transparent_40%,rgba(0,0,0,0.6)_100%)]" />
       </div>
@@ -250,7 +108,8 @@ export default async function MyPage() {
                 alt=""
                 fill
                 className="object-contain rounded-full opacity-60"
-               sizes="100vw" />
+                sizes="100vw"
+              />
             </div>
             <p className="text-arcana-muted text-lg font-serif mb-2">{t("mypage.history.empty.title", locale)}</p>
             <p className="text-arcana-muted/60 text-sm mb-6">{t("mypage.history.empty.desc", locale)}</p>
